Update README.md
Alphabet Soup, a nonprofit foundation, aims to identify applicants with the best chance of success for funding their ventures. Using historical data of over 34,000 organizations, we developed a binary classifier to predict the success of these applicants. The dataset includes metadata such as application type, affiliated sector, use case for funding, and more.